refact : webScraping 함수 모듈로 변환
- webScraping 함수가 getWeekMenu와 getTodayMenu에서 모두 쓰임
- 중복된 코드를 하나의 모듈로 빼서 사용하도록 함
